packages/services/workers/rag-worker/requirements.txt
kafka-python-ng==2.2.2
chromadb-client==0.6.3
sentence-transformers==3.4.1
